Const generics & heapless v0.7 (#81)
* Bump heapless dependency to 0.6

* Formatting

* Attempt with generic AtatCmd

* Use global const ident when deriving

* Make clippy happy

* Fix doc tests

* Fix derive recursive issue

* Address review comments

+++ atat_derive/src/parse.rs
@@ -32,7 +32,7 @@ pub struct CmdAttributes {
 pub struct ArgAttributes {
     pub value: Option<i64>,
     pub position: Option<usize>,
-    pub len: Option<u32>,
+    pub len: Option<usize>,
     pub default: bool,
 }
